extract adc to its own module
