+++ src/resources/common.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "pi")]
+pub mod adc;
+
+#[cfg(feature = "pi")]
+pub use adc::Adc;
